requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
albumentations=1.3.1=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2023.7.22=h56e8100_0
certifi=2023.5.7=py38haa95532_0
cffi=1.15.1=py38h2bbff1b_3
charset-normalizer=3.1.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
coloredlogs=15.0.1=pypi_0
contourpy=1.1.0=pypi_0
cryptography=39.0.1=py38h21b164f_0
cycler=0.11.0=pypi_0
cython=0.29.35=pypi_0
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
easydict=1.10=pypi_0
eigen=3.4.0=h2d74725_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.2.3=ha925a31_0
filelock=3.12.2=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.40.0=pypi_0
freetype=2.10.4=h546665d_1
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.10.6=nompi_he0bbb20_101
humanfriendly=10.0=pypi_0
icu=58.2=ha925a31_3
idna=3.4=py38haa95532_0
imageio=2.31.1=pypi_0
importlib-metadata=6.7.0=pyha770c72_0
importlib_metadata=6.7.0=hd8ed1ab_0
insightface=0.7.3=pypi_0
intel-openmp=2023.1.0=h57928b3_46319
ipykernel=6.15.0=pyh025b116_0
ipython=8.12.0=pyh08f2357_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py38haa95532_0
joblib=1.2.0=pypi_0
jpeg=9e=h8ffe710_2
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_core=5.3.1=py38haa244fe_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.2=pypi_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libclang=12.0.0=default_h627e005_2
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=h8ffe710_5
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=8_mkl
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.3=h0ad7f3c_0
libxslt=1.1.37=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2020.4=hb70f87d_311
mpmath=1.3.0=pypi_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
onnx=1.14.0=pypi_0
onnxruntime=1.15.0=pypi_0
opencv=4.6.0=py38hf11a4ad_3
opencv-python-headless=4.7.0.72=pypi_0
openssl=1.1.1u=h2bbff1b_0
packaging=23.1=pyhd8ed1ab_0
paho-mqtt=1.6.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h0e60522_0
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=23.1.2=py38haa95532_0
platformdirs=3.8.0=pyhd8ed1ab_0
prettytable=3.8.0=pypi_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=4.23.3=pypi_0
psutil=5.9.0=py38h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pyhd8ed1ab_0
pyopenssl=23.0.0=py38haa95532_0
pyparsing=3.0.9=pypi_0
pyreadline3=3.4.1=pypi_0
pysocks=1.7.1=py38haa95532_0
python=3.8.0=hff0d562_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=2.0.1=py3.8_cpu_0
pytorch-mutex=1.0=cpu
pywavelets=1.4.1=pypi_0
pywin32=305=py38h2bbff1b_0
pyyaml=6.0=pypi_0
pyzmq=25.1.0=py38hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h2bbfb41_5
qudida=0.0.4=pypi_0
requests=2.31.0=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
see=1.4.1=pypi_0
setuptools=67.8.0=py38haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
tk=8.6.12=h2bbff1b_0
torch=2.0.1=pypi_0
torchaudio=2.0.2+cu117=pypi_0
torchvision=0.15.2+cu117=pypi_0
tornado=6.2=py38h294d835_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.6.3=pypi_0
typing_extensions=4.7.0=pyha770c72_0
urllib3=2.0.3=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.4.2=h8cc25b3_0
zeromq=4.3.4=h0e60522_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
